Move the security banner to a new class
